requirements.txt
transformers~=4.47.0
torch~=2.6.0
overrides~=7.7.0
llama-index-core~=0.12.2
qdrant-client~=1.12.1
